ui_automation/requirements.txt
playwright>=1.44.0
pytest>=8.2.0
pytest-html>=4.1.0
allure-pytest>=2.13.0
pyyaml>=6.0.1
